Change numpy float type to string in pretrain_dataset.py [skip ci]

+++ model/framework/code/src/data/pretrain_dataset.py
@@ -12,8 +12,8 @@ class MoleculeDataset(Dataset):
         with open(smiles_path, 'r') as f:
             lines = f.readlines()
             self.smiles_list = [line.strip('\n') for line in lines]
-        self.fps = torch.from_numpy(sps.load_npz(fp_path).todense().astype(np.float32))
-        mds = np.load(md_path)['md'].astype(np.float32)
+        self.fps = torch.from_numpy(sps.load_npz(fp_path).todense().astype("float"))
+        mds = np.load(md_path)['md'].astype("float")
         mds = np.where(np.isnan(mds), 0, mds)
         self.mds = torch.from_numpy(mds)
         self.d_fps = self.fps.shape[1]
@@ -30,7 +30,7 @@ class MoleculeDataset(Dataset):
         task_pos_weights = torch.ones(self.fps.shape[1])
         num_pos = torch.sum(torch.nan_to_num(self.fps,nan=0), axis=0)
         masks = F.zerocopy_from_numpy(
-            (~np.isnan(self.fps.numpy())).astype(np.float32))
+            (~np.isnan(self.fps.numpy())).astype("float"))
         num_indices = torch.sum(masks, axis=0)
         task_pos_weights[num_pos > 0] = ((num_indices - num_pos) / num_pos)[num_pos > 0]
         return task_pos_weights
